forge_IAC_type_1/forge_agent_v1.2/requirements.txt
watchdog==3.0.0
gitpython==3.1.40
python-dotenv==1.0.0
langchain-openai==0.0.2
openai==1.3.7
google-generativeai==0.3.2
pexpect==4.8.0 
